refactor(agents): extract shared createLazyExecTool factory
