+++ src/security/dangerous-tools.ts
@@ -38,4 +38,4 @@ export const DEFAULT_GATEWAY_HTTP_TOOL_DENY = [
  * `gateway.tools.allow` can remove the default HTTP deny only for owner/trusted-operator
  * callers; non-owner identity-bearing callers must not receive server-credential wrappers.
  */
-export const GATEWAY_OWNER_ONLY_CORE_TOOLS = ["cron", "gateway", "nodes"] as const;
+export const GATEWAY_OWNER_ONLY_CORE_TOOLS = ["cron", "exec", "gateway", "nodes"] as const;
